Fixing Issue#480
